Provide a mechanism for working locally
The app previously couldn't be run without a public IP b/c it would hit
github on most requests. This override flag will allow you to work 
locally so long as a user with a given name does exist in the DB. They
still need a github record associated with them for much at all to work
but if it's set up once it'll be fine.

To pull it off, just do something like:

LOCAL_OVERRIDE_AS="Justin Searls" rails s