Adds entry-level editing
You can now edit the location of individual entries

* If an entry is set to something other than the user's default, it
appears blue
* If all of the entries on a timesheet are set to something other than
the user's default, the timesheet-level drop-down will default to that
non-default location (so at a glance you can see what they're all set 
to)


This implementation was incredibly messy relative to my initial 
expectations. In hindsight I wish I'd hacked up a post-back to the
original form instead of doing this with AJAX. Half the new code is in 
place just to keep the client & server side UI in sync.

Fixes #5