"Ready for invoice" now locks the timesheet.
Instead of a passive boolean, get confirmation and then lock the week.
This is a bit more heavy-handed than we would have liked, but folks 
going back and changing time without communicating it has caused 
several invoices to need to be resent (and worse, late). Easier to be
forced to talk about it outside the system when time is missed.

* Confirmation prompt added to make sure people know it's a one way 
street
* Added some light server-side validation to prevent the super easy
HTML hacking to the Entry class by way of an active record callback

Fixes #33