Allow admins to impersonate and set locked entries
Squeezes in the logged in user into every entry via params hacking
(Really gross way of doing it. :shrug:)

This means that the value will be null when the user creates a timesheet
but does not update it, which i think is actually meaningful for 
tracking authorship. They just hit "next" they didn't necessarily mean
to update a thing.