Dramatically revamp reminder emails
Reminder e-mails appear to just never come, and tracking it is really
hard, so here each will be bundled into a job (run by the same rake 
task as always), and each email will also be recorded for posterity.
Additionally, if an agent lacks an e-mail address in the system, it will
send an e-mail to the domain admins (finances@testdouble.com) to alert
us that their notification is not coming through.