Just ran into our first time zone issue :'-(
The heroku scheduler runs at seemingly random times, and sometimes so
late in the day that it's no longer friday UTC when it runs.

Amazingly, this seems to be the only part of the system where the 
vaguery of UTC-5 actually poses a real inconvenience since so few people
use the system after 7pm eastern on friday & saturday when things roll
over.

Hey @jasonkarns & @danthompson, how's your rails time zone-fu? If we 
change the rails config to do everything in Eastern time, will it screw
up our existing data? I'd rather the database store everything in UTC
and have the app marshal/unmarshal dates in eastern time.