Avoid running `rake db:seed` between each test
Problem was: each test relied on seeds running, which 
required fetching manually maintained fixtures over the 
harvest API. For most intents and purposes, this limited us
to a single test in our test suite. It was particularly brittle and
exercised less of the system than it could have.

Stuff we had to do:

* Extract seed script into modules that can be individually invoked 
* Look up the reference invoice week lazily and only memoize
it once it was non-nil
* Create a new Harvest Test account & related ENV/secrets
* Add to the smoke test steps to automate client/project 
creation (since the seed isn't doing it anymore)
* Use the patented @josevalim hack to ensure AR will run all queries
through a single DB connection (so the transactional strategy works)

Hard to believe it was a 5 hour yak shave to be able to say "we can run
two tests in a single process", but there we are.