remove clever defaults from DB schema
Rails really doesn't support this well and keeps breaking the default
additionally, any system that had defaults differently than what the 
test wants would break, which isn't ideal. we may as well just read the
live environment variable.

a benefit of this is that a simple config change will be able to change
the default rates of the system (not requiring a schema change); this
bit both me and @tkaufman recently